Using date filter comparator
